lock down mutations

+++ src/resolvers/Mutation.js
@@ -1,5 +1,6 @@
 import bcrypt from 'bcryptjs'
 import jwt from 'jsonwebtoken'
+import getUserId from '../utils/getUserId'
 
 const Mutation = {
 	async createUser(parent, args, { prisma }, info) {
@@ -22,19 +23,46 @@ const Mutation = {
 			token: jwt.sign({ userId: user.id }, 'thisisasecret')
 		}
 	},
-	deleteUser(parent, args, { prisma }, info) {
-		return prisma.mutation.deleteUser({ where: { id: args.id } }, info);
+	async login(parent, args, { prisma }, info) {
+		const user = await prisma.query.user({
+			where: {
+				email: args.data.email
+			}
+		})
+
+		if (!user) {
+			throw new Error('Unable to login. User or password not correct')
+		}
+
+		const isMatch = await bcrypt.compare(args.data.password, user.password);
+
+		if (!isMatch) {
+			throw new Error('Unable to login. User or password not correct')
+		}
+
+		return {
+			user,
+			token: jwt.sign({ userId: user.id }, 'thisisasecret')
+		}
+
 	},
-	updateUser(parent, args, { prisma }, info) {
+	deleteUser(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request)
+		return prisma.mutation.deleteUser({ where: { id: userId } }, info);
+	},
+	updateUser(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request)
+
 		return prisma.mutation.updateUser(
 			{
-				where: { id: args.id },
+				where: { id: userId },
 				data: args.data
 			},
 			info
 		);
 	},
-	createPost(parent, args, { prisma }, info) {
+	createPost(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request)
 		return prisma.mutation.createPost(
 			{
 				data: {
@@ -43,7 +71,7 @@ const Mutation = {
 					published: args.data.published,
 					author: {
 						connect: {
-							id: args.data.author
+							id: userId
 						}
 					}
 				}
@@ -51,7 +79,19 @@ const Mutation = {
 			info
 		);
 	},
-	deletePost(parent, args, { prisma }, info) {
+	async deletePost(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request)
+		const postExists = await prisma.exists.Post({
+			id: args.id,
+			author: {
+				id: userId
+			}
+		})
+
+		if (!postExists) {
+			throw new Error('Unable to delete post')
+		}
+
 		return prisma.mutation.deletePost(
 			{
 				where: {
@@ -61,7 +101,19 @@ const Mutation = {
 			info
 		);
 	},
-	updatePost(parent, args, { prisma }, info) {
+	async updatePost(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request)
+		const postExists = await prisma.exists.Post({
+			id: args.id,
+			author: {
+				id: userId
+			}
+		})
+
+		if(!postExists){
+			throw new Error('Unable to update post!')
+		}
+
 		return prisma.mutation.updatePost(
 			{
 				where: {
@@ -72,13 +124,15 @@ const Mutation = {
 			info
 		);
 	},
-	createComment(parent, args, { prisma }, info) {
+	createComment(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request)
+
 		return prisma.mutation.createComment(
 			{
 				data: {
 					text: args.data.text,
 					author: {
-						connect: { id: args.data.author }
+						connect: { id: userId }
 					},
 					post: {
 						connect: { id: args.data.post }
@@ -88,14 +142,35 @@ const Mutation = {
 			info
 		);
 	},
-	deleteComment(parent, args, { prisma }, info) {
+	async deleteComment(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request)
+		const commentExists = await prisma.exists.Comment({
+			id: args.id,
+			author: {
+				id: userId
+			}
+		})
+		if(!commentExists){
+			throw new Error('Unable to delete comment')
+		}
+
 		return prisma.mutation.deleteComment({
 			where: {
 				id: args.id
 			}
 		}, info)
 	},
-	updateComment(parent, args, { db, pubsub }, info) {
+	async updateComment(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request)
+		const commentExists = await prisma.exists.Comment({
+			id: args.id,
+			author: {
+				id: userId
+			}
+		})
+		if(!commentExists){
+			throw new Error('Unable to update comment')
+		}
 		return prisma.mutation.updateComment({
 			where: {
 				id: args.id
